restructured and updated the package
