--- telemetry/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	tel "github.com/chef/go-libs/telemetry/telemetry"
-)
-
-func main() {
-	tel.Setup()
-}
