+++ telemetry/sender.go
@@ -1,1 +1,1 @@
-package telemeter
+package telemetry
