+++ telemetry/telemetry.go
@@ -1,1 +1,1 @@
-package main
+package telemetry
